添加支付宝支付模块

+++ src/main/java/com/lab/hosaily/core/client/dao/mapper/PaymentMapper.java
@@ -38,5 +38,5 @@ public interface PaymentMapper{
     /**
      * 查询金额
      */
-    Long priceByPurchaseId(@Param("params") Criteria criteria);
+    Double priceByPurchaseId(@Param("params") Criteria criteria);
 }
