+++ src/main/java/com/lab/hosaily/core/client/service/PaymentService.java
@@ -5,5 +5,5 @@ public interface PaymentService{
     /**
      * 根据购买记录统计已成功付款金额
      */
-    long priceWechatMerchantPayByPurchaseId(String purchaseId);
+    double priceWechatMerchantPayByPurchaseId(String purchaseId);
 }
